+++ packages/block-editor/src/hooks/index.js
@@ -22,6 +22,7 @@ import duotone from './duotone';
 import fontFamily from './font-family';
 import fontSize from './font-size';
 import textAlign from './text-align';
+import textShadow from './text-shadow';
 import fitText from './fit-text';
 import border from './border';
 import customCSS from './custom-css';
@@ -68,6 +69,7 @@ createBlockListBlockFilter( [
 	duotone,
 	fontFamily,
 	fontSize,
+	textShadow,
 	fitText,
 	border,
 	customCSS,
@@ -88,6 +90,7 @@ createBlockSaveFilter( [
 	style,
 	fontFamily,
 	fontSize,
+	textShadow,
 ] );
 
 export { useCustomSides } from './dimensions';