+++ packages/blocks/src/store/test/private-selectors.js
@@ -39,6 +39,7 @@ describe( 'private selectors', () => {
 				'wideSize',
 				'blockGap',
 				'textAlign',
+				'textShadow',
 			] );
 		} );
 
@@ -64,6 +65,7 @@ describe( 'private selectors', () => {
 				'blockGap',
 				'textAlign',
 				'textDecoration',
+				'textShadow',
 			] );
 		} );
 
@@ -92,6 +94,7 @@ describe( 'private selectors', () => {
 				'blockGap',
 				'textAlign',
 				'textTransform',
+				'textShadow',
 				'letterSpacing',
 			] );
 		} );
@@ -141,6 +144,7 @@ describe( 'private selectors', () => {
 				'blockGap',
 				'textAlign',
 				'textTransform',
+				'textShadow',
 				'letterSpacing',
 			] );
 		} );