+++ packages/grid/src/dashboard-grid/grid-item.tsx
@@ -29,6 +29,8 @@ export function GridItem( {
 	item,
 	maxColumns,
 	disabled = false,
+	draggable = true,
+	resizable = true,
 	verticalResizable = true,
 	interacting = false,
 	dragging = false,
@@ -52,6 +54,8 @@ export function GridItem( {
 	} | null >( null );
 	const itemRef = useRef< HTMLDivElement >( null );
 	const contentRef = useRef< HTMLDivElement >( null );
+	const dragDisabled = disabled || ! draggable;
+	const resizeDisabled = disabled || ! resizable;
 	const {
 		attributes,
 		listeners,
@@ -60,7 +64,7 @@ export function GridItem( {
 		isDragging,
 	} = useSortable( {
 		id: item.key,
-		disabled,
+		disabled: dragDisabled,
 	} );
 	const mergedRef = useMergeRefs( [ itemRef, setNodeRef ] );
 	const contentMergedRef = useMergeRefs( [ contentRef ] );
@@ -166,7 +170,7 @@ export function GridItem( {
 				{ ...listeners }
 				style={ {
 					height: '100%',
-					cursor: getItemCursor( disabled, interacting ),
+					cursor: getItemCursor( dragDisabled, interacting ),
 				} }
 			>
 				<div
@@ -175,7 +179,7 @@ export function GridItem( {
 					style={ continuousContentStyle }
 				>
 					{ children }
-					{ ! disabled && (
+					{ ! resizeDisabled && (
 						<ResizeHandle
 							itemId={ item.key }
 							verticalResizable={ verticalResizable }