+++ packages/reusable-blocks/src/components/reusable-blocks-menu-items/index.js
@@ -1,28 +1,13 @@
-import { BlockSettingsMenuControls } from '@wordpress/block-editor';
 import deprecated from '@wordpress/deprecated';
-import ReusableBlockConvertButton from './reusable-block-convert-button';
-import ReusableBlocksManageButton from './reusable-blocks-manage-button';
 
-export default function ReusableBlocksMenuItems( { rootClientId } ) {
+/**
+ * Deprecated no-op ReusableBlocksMenuItems component.
+ *
+ * @return {null} Nothing.
+ */
+export default function ReusableBlocksMenuItems() {
 	deprecated( 'wp.reusableBlocks.ReusableBlocksMenuItems', {
 		since: '7.1',
 	} );
-	return (
-		<BlockSettingsMenuControls>
-			{ ( { onClose, selectedClientIds } ) => (
-				<>
-					<ReusableBlockConvertButton
-						clientIds={ selectedClientIds }
-						rootClientId={ rootClientId }
-						onClose={ onClose }
-					/>
-					{ selectedClientIds.length === 1 && (
-						<ReusableBlocksManageButton
-							clientId={ selectedClientIds[ 0 ] }
-						/>
-					) }
-				</>
-			) }
-		</BlockSettingsMenuControls>
-	);
+	return null;
 }