+++ packages/ui/src/form/primitives/select/popup.tsx
@@ -4,7 +4,6 @@ import { forwardRef } from '@wordpress/element';
 import { Portal } from './portal';
 import { Positioner } from './positioner';
 import { renderSlotWithChildren } from '../../../utils/render-slot-with-children';
-import { ThemeProvider } from '../../../utils/theme-provider';
 import itemPopupStyles from '../../../utils/css/item-popup.module.css';
 import type { SelectPopupProps } from './types';
 
@@ -14,23 +13,21 @@ export const Popup = forwardRef< HTMLDivElement, SelectPopupProps >(
 		ref
 	) {
 		const popupContent = (
-			<ThemeProvider>
-				<_Select.Popup
-					ref={ ref }
-					className={ clsx( itemPopupStyles.popup, className ) }
-					{ ...restProps }
-				>
-					<_Select.List className={ itemPopupStyles.list }>
-						<div
-							className={
-								itemPopupStyles[ 'list-scrollable-container' ]
-							}
-						>
-							{ children }
-						</div>
-					</_Select.List>
-				</_Select.Popup>
-			</ThemeProvider>
+			<_Select.Popup
+				ref={ ref }
+				className={ clsx( itemPopupStyles.popup, className ) }
+				{ ...restProps }
+			>
+				<_Select.List className={ itemPopupStyles.list }>
+					<div
+						className={
+							itemPopupStyles[ 'list-scrollable-container' ]
+						}
+					>
+						{ children }
+					</div>
+				</_Select.List>
+			</_Select.Popup>
 		);
 
 		const positionedPopup = renderSlotWithChildren(